outline of a standardized BaseGame cls
