+++ raycast/main.py
@@ -1,136 +1,16 @@
-
-############## game.py ##############
-# TODO web exports only support a single src file right now (I think)... so everything is slapped into one mega file
-
+import math
+import random
 
 import katagames_sdk.engine as kataen
 
-import time
-import collections
+from BaseGame import BaseGame as Game
+from Vector2 import Vector2
+
 
 pygame = kataen.import_pygame()
 EventReceiver = kataen.EventReceiver
 EngineEvTypes = kataen.EngineEvTypes
 
-
-class Game:
-    """Base class for games."""
-
-    def __init__(self, track_fps=True):
-        self._fps_n_frames = 10 if track_fps else 0
-        self._fps_tracker_logic = collections.deque()
-        self._fps_tracker_rendering = collections.deque()
-        self._tick = 0
-
-        self._cached_info_text = None
-        self._info_font = None
-
-    def start(self):
-        """Starts the game loop. This method will not exit until the game has finished execution."""
-        kataen.init(self._get_mode_internal())
-
-        li_recv = [kataen.get_game_ctrl(), self.build_controller()]
-        for recv_obj in li_recv:
-            recv_obj.turn_on()
-
-        self.pre_update()
-
-        li_recv[0].loop()
-        kataen.cleanup()
-
-    def get_mode(self) -> str:
-        """returns: "HD', 'OLD_SCHOOL', or 'SUPER_RETRO'"""
-        return 'OLD_SCHOOL'
-
-    def is_running_in_web(self) -> bool:
-        return kataen.runs_in_web()
-
-    def get_screen_size(self):
-        return kataen.get_screen().get_size()
-
-    def get_tick(self) -> int:
-        return self._tick
-
-    def pre_update(self):
-        pass
-
-    def render(self, screen):
-        raise NotImplementedError()
-
-    def update(self, events, dt):
-        raise NotImplementedError()
-
-    def render_text(self, screen, text, size=12, pos=(0, 0), xanchor=0, color=(255, 255, 255), bg_color=None):
-        if self._info_font is None or self._info_font.get_height() != size:
-            self._info_font = pygame.font.Font(None, size)
-        lines = text.split("\n")
-        y = pos[1]
-        for l in lines:
-            surf = self._info_font.render(l, True, color, bg_color)
-            screen.blit(surf, (int(pos[0] - xanchor * surf.get_width()), y))
-            y += surf.get_height()
-
-    def get_fps(self, logical=True) -> float:
-        q = self._fps_tracker_logic if logical else self._fps_tracker_rendering
-        if len(q) <= 1:
-            return 0
-        else:
-            total_time_secs = q[-1] - q[0]
-            n_frames = len(q)
-            if total_time_secs <= 0:
-                return float('inf')
-            else:
-                return (n_frames - 1) / total_time_secs
-
-    def _render_internal(self, screen):
-        if self._fps_n_frames > 0:
-            self._fps_tracker_rendering.append(time.time())
-            if len(self._fps_tracker_rendering) > self._fps_n_frames:
-                self._fps_tracker_rendering.popleft()
-        self.render(screen)
-
-    def _update_internal(self, events, dt):
-        if self._fps_n_frames > 0:
-            self._fps_tracker_logic.append(time.time())
-            if len(self._fps_tracker_logic) > self._fps_n_frames:
-                self._fps_tracker_logic.popleft()
-        self.update(events, dt)
-        self._tick += 1
-
-    def _get_mode_internal(self):
-        mode_str = self.get_mode().upper()
-        if mode_str == 'HD':
-            return kataen.HD_MODE
-        elif mode_str == 'OLD_SCHOOL':
-            return kataen.OLD_SCHOOL_MODE
-        elif mode_str == 'SUPER_RETRO':
-            return kataen.SUPER_RETRO_MODE
-        else:
-            raise ValueError("Unrecognized mode: {}".format(mode_str))
-
-    class _GameViewController(EventReceiver):
-        def __init__(self, game):
-            super().__init__()
-            self._game = game
-            self._event_queue = []
-            self._last_update_time = time.time()
-
-        def proc_event(self, ev, source):
-            if ev.type == EngineEvTypes.PAINT:
-                self._game._render_internal(ev.screen)
-            elif ev.type == EngineEvTypes.LOGICUPDATE:
-                cur_time = time.time()
-                self._game._update_internal(self._event_queue, cur_time - self._last_update_time)
-                self._last_update_time = cur_time
-                self._event_queue.clear()
-            else:
-                self._event_queue.append(ev)
-
-    def build_controller(self) -> EventReceiver:
-        return Game._GameViewController(self)
-
-
-############## game.py ##############
 
 ############## art.py ##############
 
@@ -156,111 +36,6 @@ class Art:
 ############## art.py ##############
 
 ############## raycaster.py ##############
-
-import math
-import random
-
-
-class Vector2:
-    # TODO pygame.Vector2 doesn't seem to be supported yet. So I made my own >:(
-
-    def __init__(self, x, y=0.0):
-        if isinstance(x, Vector2):
-            self.x = x.x
-            self.y = x.y
-        else:
-            self.x = x
-            self.y = y
-
-    def __getitem__(self, idx):
-        if idx == 0:
-            return self.x
-        else:
-            return self.y
-
-    def __len__(self):
-        return 2
-
-    def __iter__(self):
-        return (v for v in (self.x, self.y))
-
-    def __add__(self, other: 'Vector2'):
-        return Vector2(self.x + other[0], self.y + other[1])
-
-    def __sub__(self, other: 'Vector2'):
-        return Vector2(self.x - other[0], self.y - other[1])
-
-    def __mul__(self, other: float):
-        return Vector2(self.x * other, self.y * other)
-
-    def __neg__(self):
-        return Vector2(-self.x, -self.y)
-
-    def __eq__(self, other: 'Vector2'):
-        return self.x == other[0] and self.y == other[1]
-
-    def __hash__(self):
-        return hash((self.x, self.y))
-
-    def __repr__(self):
-        return "Vector2({}, {})".format(self.x, self.y)
-
-    def dot(self, other):
-        return self.x * other[0] + self.y * other[1]
-
-    def rotate_ip(self, degrees):
-        theta = math.radians(degrees)
-        cs = math.cos(theta)
-        sn = math.sin(theta)
-        x = self.x * cs - self.y * sn
-        y = self.x * sn + self.y * cs
-        self.x = x
-        self.y = y
-
-    def rotate(self, degrees):
-        res = Vector2(self)
-        res.rotate_ip(degrees)
-        return res
-
-    def to_ints(self):
-        return Vector2(int(self.x), int(self.y))
-
-    def length(self):
-        return math.sqrt(self.x * self.x + self.y * self.y)
-
-    def length_squared(self):
-        return self.x * self.x + self.y * self.y
-
-    def scale_to_length(self, length, or_else_rand=True) -> 'Vector2':
-        cur_length = self.length()
-        if cur_length == 0 and length != 0:
-            if or_else_rand:
-                res = Vector2(0, length)
-                res.rotate_ip(360 * random.random())
-                return res
-            else:
-                raise ValueError("Cannot scale vector with length 0")
-        else:
-            mult = length / cur_length
-            return self * mult
-
-    def angle_to(self, other):
-        dot = self.dot(other)
-        mags = self.length() * math.sqrt(other[0] * other[0] + other[1] * other[1])
-        if mags > 0.001:
-            neg_dot = dot < 0
-            dot = abs(dot)
-            det = dot / mags
-            if det <= 0.999:
-                if neg_dot:
-                    return 180 - math.degrees(math.acos(det))
-                else:
-                    return math.degrees(math.acos(det))
-        return 0
-
-    def distance_to(self, other):
-        return (self - other).length()
-
 
 class RayEmitter:
